+++ packages/winnow/test/aggregate.js
@@ -88,6 +88,24 @@ test('Use multiple group bys', t => {
   t.ok(results[0].avg_Trunk_Diameter)
 })
 
+test('Use multiple group bys ESRI style', t => {
+  t.plan(4)
+  const options = {
+    aggregates: [
+      {
+        type: 'avg',
+        field: 'Trunk_Diameter'
+      }
+    ],
+    groupByFieldsForStatistics: 'Genus,Common_Name'
+  }
+  const results = winnow.query(features, options)
+  t.equal(results.length, 310)
+  t.ok(results[0].Genus)
+  t.ok(results[0].Common_Name)
+  t.ok(results[0].avg_Trunk_Diameter)
+})
+
 test('Use a group by with a where clause', t => {
   t.plan(3)
   const options = {